+++ java/src/com/protobytes/buffers/core/BuffInt.java
@@ -3,7 +3,7 @@ package com.protobytes.buffers.core;
 import com.protobytes.buffers.BuffType;
 
 /**
- * 4个字节，介于 -2147483648 和 2147483647 之间的 32 位带符号整数。
+ * A 32-bit signed integer between -2147483648 and 2147483647.
  * 
  * @author Agua.L
  *
@@ -17,7 +17,7 @@ public class BuffInt extends BuffBase {
 
 	public void setValue(Object value) {
 		this.value = (Integer) value;
-		if (this.value.intValue() < Integer.MIN_VALUE || this.value.intValue() > Integer.MAX_VALUE)
+		if (this.value < Integer.MIN_VALUE || this.value > Integer.MAX_VALUE)
 			System.err.println("BuffInt 的值超出了 -214748364 8到 2147483647 的范围");
 	}
